test: 통합 테스트 추가 및 불필요한 테스트 삭제
